Add a Storage Drives page (#10)
This PR adds a GUI page for unmounting USB drives (which are
auto-mounted by https://github.com/openUC2/rpi-imswitch-os/pull/44) and
for mounting unmounted USB drives. This can also be considered a
prototype in preparation for making an HTTP JSON API for
programamtically unmounting USB drives.

This work is tracked on Notion at
https://www.notion.so/Support-hotplugging-of-USB-drives-2a74e612c78a808c999ad81fb54a7c27?source=copy_link